Factored in some lessons learned from Mantl
The Mantl terraform configuration is implemented as a "modules"
directory of libraries that each focus on a distinct functional segment
of AWS, and a main *.tf file that is a sibling of the library module
root dir.  The main file calls into the libraries and deals with overall
flow and sets variables to define which regions/availability zones DC/OS
is to be deployed in and manages the required degree of repetition in
order to provision the desired number of EC2 nodes.

The pre-existing dcos.tf file is still here for reference.  It has not
been editted.  It pre-dated the content under modules, and so uses none
of it.

newOp.tf is the newly-created root terraform file.  It should be
functional still, except for a portion attempting to deal with Route 53
that I had commented out due to problems that had keep it from working
corectly.